* Fixed memory corruption bug in Analyzer core module that could crash the system on
  non-power-of-two buffer sizes. Affected plugins: Parametric Equalizer, Graphic Equalizer,
  Spectrum Analyzer, Multiband Compressor.

+++ src/core/util/Counter.cpp
@@ -71,7 +71,7 @@ namespace lsp
 
     bool Counter::submit(size_t samples)
     {
-        ssize_t left    = ssize_t(nCurrent) - samples;
+        ssize_t left    = ssize_t(nCurrent) - ssize_t(samples);
         if (left <= 0)
         {
             nCurrent        = nInitial + (left % ssize_t(nInitial));
